# Another Compiler Refactor: Performance & Cleanup