@ferd 2020-05-13
Benchmark with last stable rebar3 (3.13.2), which had a faster initial build, but was less good at increments:
- `./rebar3 compile:  136.16s user 33.45s system 184% cpu 1:28.75 total` 
- `./rebar3 as test compile:  153.52s user 37.93s system 182% cpu 1:45.16 total`

The current branch is therefore ~20x faster than last stable release on a larger sparse repo for basic incremental builds. Obviously these improvements aren't as visible on regular, smaller repos, but that's still significant since we don't have significant regressions there either.

That's a decent boost for the smaller price to pay on the first build.